requirements.txt
APScheduler==3.6.3
asgiref==3.5.0
beautifulsoup4==4.10.0
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==4.0.2
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-api-key==2.1.0
git-filter-repo==2.34.0
gunicorn==20.1.0
idna==3.3
numpy==1.22.2
pandas==1.4.1
Pillow==9.0.1
psutil==5.9.0
psycopg2==2.9.3
python-dateutil==2.8.2
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
whitenoise==6.0.0
xmltodict==0.12.0
